requirements.txt
requests>=2.32.0
pandas>=2.2.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
